refactor: (#35) App 레이어에서 Axios 인터셉터 설정하도록 작성
